% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recovery-package.R
\docType{package}
\name{Recovery-package}
\alias{Recovery}
\alias{Recovery-package}
\title{Recovery: Recover Missing Diagnosis Test Data With Constraint-Solving}
\description{
Recovery provides an interface to constraint solvers aimed at missing diagnostic test data recovery.
}
\author{
\strong{Maintainer}: Raoul Schorer \email{raoul.schorer@hug.ch} (\href{https://orcid.org/0000-0001-7848-7021}{ORCID})

}
\keyword{internal}
